Delete DocC workflow for building and publishing Swift documentation on GitHub Pages.
The commit removes the `.github/workflows/docc.yml` file, which contained a workflow for building Swift DocC and publishing it on GitHub Pages. The workflow was triggered by any push to any branch. It set up the necessary permissions and environment variables, checked out the code, built the documentation using SwiftPM, and committed and pushed the generated documentation to the `gh-pages` branch.

This change is being made because we no longer need this specific workflow for our project.