Documentation building via github action (#4)
* feat: Add workflow to publish DocC documentation

This commit adds a new workflow file `publish-docc.yml` that enables the publishing of DocC documentation. The workflow is triggered on release events and manual dispatches. It runs on macOS 13 and performs the following steps:
- Checks out the repository
- Generates DocC documentation using Xcodebuild
- Deploys the generated documentation to Github Pages

The deployment is done using the `github-pages-deploy-action@v4.5.0` action, which pushes the documentation to a separate branch named "docc". The commit message for each deployment includes the updated version number from the release event.

* Add support for publishing DocC documentation on the 'Documentation' branch.

This commit adds a new trigger to the workflow file `publish-docc.yml` that allows publishing DocC documentation when changes are pushed to the 'Documentation' branch.

* Update minimum Swift version to 5.9

The code change updates the minimum required version of Swift in the Package.swift file from 5.10 to 5.9. This ensures compatibility with older versions of Swift when building the package.

* Update Xcode version and macOS environment in publish-docc.yml

The commit updates the Xcode version to 15.3 and the macOS environment to 14 in the publish-docc.yml file.

* Update Swift tools version to 5.10

- Updated the Swift tools version in Package.swift from 5.9 to 5.10.

Update Documentation links

- Updated the documentation links in README.md to point to the new location.

Remove specific references

- Removed specific references to Kiwi.com in Bones.md and GettingStarted.md files.

* Update publish-docc.yml workflow

- Removed push event trigger for 'Documentation' branch
- Workflow now triggered only by release and manual dispatch events

* Updated .gitignore and removed unnecessary IDE files

The .gitignore file has been updated to ignore all files in the .idea/ directory. This change will prevent any user-specific IDE settings from being committed to the repository. Additionally, several unnecessary IDE-related files have been deleted from the repository. These changes help keep the repository clean and focused on source code only.

* Renamed and updated documentation articles

The documentation articles have been renamed and updated for better clarity. The 'Tokens' article has been renamed to 'Foundation', while the 'Components' article is now called 'Supports'. Changes in the code include refactoring of shadow application on colors, addition of a new test view, and removal of unnecessary comments from Separator component.

* Fixed shadow function syntax in documentation

Corrected the syntax for the shadow function in the Foundation.md file of the Documentation. The closing parentheses were missing, causing potential confusion for users trying to implement this feature. This commit ensures that all code examples are accurate and easy to understand.